fix(bruno-request): add tests for URL handling in pre-request and post-response scripts
